Add ticket logging

+++ commands/ticket-close.js
@@ -1,6 +1,8 @@
 const fs = require('fs');
+const { exec } = require('child_process');
 const { SlashCommandBuilder } = require('@discordjs/builders');
-const { botLogs, ticketsChannel, ticketsRole, ticketsCategory } = require('./commands_config.json');
+const { botLogs, ticketsChannel, ticketsRole, ticketsCategory, ticketsLogs } = require('./commands_config.json');
+const { token } = require('../config.json');
 
 module.exports = {
 	data: new SlashCommandBuilder()
@@ -32,6 +34,27 @@ module.exports = {
 				return console.error(err);
 			}
 		});
-        interaction.channel.delete();
+        const command = `dotnet /root/DiscordChatExporter/DiscordChatExporter.Cli.dll export -t ${token} -c ${interaction.channel.id} -o ../ticket-logs/${interaction.channel.name}.html`;
+        exec(command,
+    function (error, stdout, stderr) {
+        // console.log('stdout: ' + stdout);
+        // console.log('stderr: ' + stderr);
+        if (error !== null) {
+             console.log('exec error: ' + error);
+        }
+        else {
+            interaction.guild.channels.fetch(ticketsLogs).then(ticketLogsChannel => 
+            {
+                ticketLogsChannel.send({
+                    files: [`../ticket-logs/${interaction.channel.name}.html`]
+                });
+            });
+            interaction.channel.delete();
+        }
+    });
+    interaction.reply({
+        content: 'A dar backup dos dados e a fechar o ticket. Isto pode demorar um bocado...',
+        ephemeral: true,
+    });  
 	},
 };
